tests/xtimer_benchmarks: Remove test
As xtimer is now obsolete we can remove this.
There were also some issues in the CI and this seems the most approprite way of resolving the issue.
