--- tests/xtimer_benchmarks/Makefile
@@ -1,16 +1,0 @@
-include ../Makefile.tests_common
-
-FEATURES_REQUIRED += periph_gpio
-FEATURES_REQUIRED += HIL_DUT_IC_PIN
-FEATURES_REQUIRED += HIL_DUT_IC_PORT
-
-USEMODULE += shell
-USEMODULE += random
-USEMODULE += xtimer
-
-USE_JSON_SHELL_PARSER ?= 1
-ROBOT_HAS_PLOTS ?= 1
-
-CFLAGS+= -DTIMER_TRACE_PIN=$(HIL_DUT_IC_PIN) -DTIMER_TRACE_PORT=$(HIL_DUT_IC_PORT)
-
-include $(RIOTBASE)/Makefile.include
