--- tests/ztimer_benchmarks/tests
@@ -1,1 +1,0 @@
-../xtimer_benchmarks/tests/
